backend/pyproject.toml
[project]
name = "ect-backend"
version = "1.0.0"
description = "ECT Shift Assignment Platform - FastAPI Backend"
requires-python = ">=3.12"
dependencies = [
    # FastAPI & Server
    "fastapi>=0.133.1",
    "uvicorn[standard]==0.27.0",
    "python-multipart>=0.0.22",
    # Data processing
    "pandas==2.2.0",
    "numpy==1.26.3",
    # Utilities
    "python-dotenv==1.0.0",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "httpx==0.26.0",
    # HTML generation
    "jinja2>=3.1.6",
    # Date utilities
    "python-dateutil==2.8.2",
    # Google API
    "google-auth==2.27.0",
    "google-auth-oauthlib==1.2.0",
    "google-api-python-client==2.116.0",
    # Security
    "cryptography>=46.0.5",
    "slowapi~=0.1.9",
    "PyJWT~=2.10",
    # Firebase Authentication (for phone OTP)
    "firebase-admin>=6.5.0",
    # PostgreSQL (for cloud deployment with Neon)
    "psycopg2-binary>=2.9",
]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "flake8>=6.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
